Add automatic linter
To make the code look better, apply an automatic linter. The chosen linter is the
Google-java-formatter from https://github.com/google/google-java-format.
It is included as a https://pre-commit.com/ rule, so that it can be automatically
applied whenever the code is committed. Additionally, a github-action makes sure
that all code is correctly linted.

This PR is huge, because it has the linter applied to the code.

Closes #223