Shuffle keys for /v1/gaen/getExposed endpoint
Up to now for the /v1/gaen/getExposed endpoint, the order of the keys returned to the
clients has been the same as the order of submitting the keys. This reveals some
information that might be used to detect if somebody submitted a valid key or not.

Even if it's already fixed in /v2/gaen/exposed endpoint, this PR also shuffles
the keys for the /v1 endpoint.

Closes #274